Add the All Type (this is so powerful :star2: )
Programmatically find all converters

Replace Null type and converter by General
