+++ lib/poleica/converters/general.rb
@@ -1,8 +1,13 @@
 # -*- encoding: utf-8 -*-
 module Poleica
   module Converters
-    # The General converter, it handles any file type.
+    # The General converter, it handles any file type and return null for
+    # unknown conversion
     class General
+
+      COMPATIBLE_TYPES = [
+        Types::All
+      ]
 
       attr_reader :polei
 
@@ -12,6 +17,10 @@ module Poleica
 
       def to_bin(options = {})
         File.binread(polei.path)
+      end
+
+      def method_missing(method, *args, &block)
+        super unless method =~ /^to_/
       end
     end # class General
   end # module Converters
