+++ lib/poleica/types/general.rb
@@ -3,10 +3,14 @@ module Poleica
   module Types
     # Null Object Pattern Type, this is the type returned if no compatible
     # types are found
-    class Null
+    class General
+
+      COMPATIBLE_MIMETYPES  = []
+      COMPATIBLE_EXTENSIONS = []
+
       def initialize(file_path)
 
       end
-    end # class Null
+    end # class General
   end # module Types
 end # module Poleica
