+++ test/poleica/converters/general_test.rb
@@ -8,4 +8,18 @@ class GeneralTest < Minitest::Test
     gif_polei = Poleica::Polei.new(gif_path)
     assert_equal(gif_data, gif_polei.to_bin)
   end
+
+  def test_it_returns_nil_on_to_method
+    method_name = :to_random
+    file        = Minitest::Mock.new
+    converter   = Poleica::Converters::General.new(file)
+    assert_nil(converter.send(method_name))
+  end
+
+  def test_it_raise_no_method_error_on_other_calls
+    method_name = :random
+    file        = Minitest::Mock.new
+    converter   = Poleica::Converters::General.new(file)
+    assert_raises(NoMethodError) { converter.send(method_name) }
+  end
 end # class GeneralTest
