--- test/poleica/converters/null_test.rb
@@ -1,18 +1,0 @@
-# -*- encoding: utf-8 -*-
-require 'test_helper'
-# Test the Null Converter Module
-class NullTest < Minitest::Test
-  def test_it_returns_nil_on_to_method
-    method_name = :to_random
-    file        = Minitest::Mock.new
-    converter   = Poleica::Converters::Null.new(file)
-    assert_nil(converter.send(method_name))
-  end
-
-  def test_it_raise_no_method_error_on_other_calls
-    method_name = :random
-    file        = Minitest::Mock.new
-    converter   = Poleica::Converters::Null.new(file)
-    assert_raises(NoMethodError) { converter.send(method_name) }
-  end
-end # class NullTest
